Read CFD and Forex tick timestamps as double intead int
This allows resolution higher than millisecond.

+++ Common/Data/Market/Tick.cs
@@ -247,7 +247,7 @@ namespace QuantConnect.Data.Market
                         var csv = line.ToCsv(3);
                         Symbol = config.Symbol;
                         TickType = TickType.Quote;
-                        Time = date.Date.AddMilliseconds(csv[0].ToInt64()).ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
+                        Time = date.Date.AddMilliseconds((double)csv[0].ToDecimal()).ConvertTo(config.DataTimeZone, config.ExchangeTimeZone);
                         BidPrice = csv[1].ToDecimal();
                         AskPrice = csv[2].ToDecimal();
                         Value = (BidPrice + AskPrice) / 2;
